% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Shiny app}
\usage{
launch_app()
}
\description{
Launch Shiny app to show plot comparison for grassland bird population and production rate of livestock and crops in North America.
}
